// src/generated/api/payments/payments.schemas.ts
export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PaymentMethod = {
  BANK_TRANSFER: 'BANK_TRANSFER',
  COD: 'COD',
  VNPAY: 'VNPAY',
} as const;

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PaymentStatus = {
  PENDING: 'PENDING',
  AWAITING_CONFIRMATION: 'AWAITING_CONFIRMATION',
  NEED_REVIEW: 'NEED_REVIEW',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED',
  REFUNDED: 'REFUNDED',
} as const;
// ...
export type PaymentEvidenceStatus =
  (typeof PaymentEvidenceStatus)[keyof typeof PaymentEvidenceStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PaymentEvidenceStatus = {
  PENDING_REVIEW: 'PENDING_REVIEW',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED',
  CANCELLED: 'CANCELLED',
} as const;